refactor(web): extract session target picker from home page
Follow-up committed on #919: move SessionTarget state, option building,
labels, branch/multi-repo handling, and request-field construction into
useSessionTargetPicker, and the target/branch/multi-select controls into
a SessionTargetPicker component. The page keeps model, prompt, and
warming. Behavior-preserving; covered by the existing Home tests.
